pyproject.toml
[tool.poetry]
name = "info-collector"
version = "0.1.0"
description = ""
authors = ["bbrfkr <bbrfkr@gmail.com>"]
readme = "README.md"
packages = [{include = "info_collector"}]

[tool.poetry.dependencies]
python = "^3.9"
redis = "^4.5.4"
feedparser = "^6.0.10"
pyyaml = "^6.0"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
